+++ ellipse-core/src/main/java/com/tomcz/ellipse/common/CoroutineScopeExtension.kt
@@ -2,7 +2,6 @@ package com.tomcz.ellipse.common
 
 import com.tomcz.ellipse.PartialState
 import com.tomcz.ellipse.Processor
-import com.tomcz.ellipse.internal.FlowEffectProcessor
 import com.tomcz.ellipse.internal.FlowProcessor
 import com.tomcz.ellipse.internal.util.consume
 import kotlinx.coroutines.CoroutineScope
@@ -10,15 +9,6 @@ import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
 import kotlinx.coroutines.launch
-
-fun <EV : Any, EF : Any> CoroutineScope.processor(
-    prepare: suspend EllipseContext<Nothing, EF>.() -> Unit = {},
-    onEvent: suspend EllipseContext<Nothing, EF>.(EV) -> Unit = {},
-): Processor<EV, Nothing, EF> = FlowEffectProcessor(
-    scope = this,
-    prepare = prepare,
-    onEvent = onEvent
-)
 
 fun <EV : Any, ST : Any, PA : PartialState<ST>, EF : Any> CoroutineScope.processor(
     initialState: ST,
@@ -31,6 +21,16 @@ fun <EV : Any, ST : Any, PA : PartialState<ST>, EF : Any> CoroutineScope.process
     onEvent = onEvent
 )
 
+fun <EV : Any, EF : Any> CoroutineScope.processor(
+    prepare: suspend EllipseContext<Unit, EF>.() -> Unit = {},
+    onEvent: suspend EllipseContext<Unit, EF>.(EV) -> Unit = {},
+): Processor<EV, Unit, EF> = FlowProcessor(
+    scope = this,
+    initialState = Unit,
+    prepare = { prepare(); emptyFlow() },
+    onEvent = { onEvent(it); emptyFlow() }
+)
+
 @FlowPreview
 fun <EV : Any, ST : Any, EF : Any> CoroutineScope.onProcessor(
     processor: () -> Processor<EV, ST, EF>,
@@ -41,19 +41,6 @@ fun <EV : Any, ST : Any, EF : Any> CoroutineScope.onProcessor(
     consume(
         processor = processor(),
         render = onState,
-        trigger = onEffect,
-        viewEvents = viewEvents()
-    )
-}
-
-@FlowPreview
-fun <EV : Any, EF : Any> CoroutineScope.onProcessor(
-    processor: () -> Processor<EV, Nothing, EF>,
-    viewEvents: () -> List<Flow<EV>> = { emptyList() },
-    onEffect: (EF) -> Unit = {},
-) = launch {
-    consume(
-        processor = processor(),
         trigger = onEffect,
         viewEvents = viewEvents()
     )
