+++ ellipse-core/src/main/java/com/tomcz/ellipse/common/ViewModelExtension.kt
@@ -7,14 +7,6 @@ import com.tomcz.ellipse.Processor
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
 
-fun <EV : Any, EF : Any> ViewModel.processor(
-    prepare: suspend EllipseContext<Nothing, EF>.() -> Unit = {},
-    onEvent: suspend EllipseContext<Nothing, EF>.(EV) -> Unit = {},
-): Processor<EV, Nothing, EF> = viewModelScope.processor(
-    prepare = prepare,
-    onEvent = onEvent
-)
-
 fun <EV : Any, ST : Any, PA : PartialState<ST>, EF : Any> ViewModel.processor(
     initialState: ST,
     prepare: suspend EllipseContext<ST, EF>.() -> Flow<PA> = { emptyFlow() },
@@ -23,4 +15,13 @@ fun <EV : Any, ST : Any, PA : PartialState<ST>, EF : Any> ViewModel.processor(
     initialState = initialState,
     prepare = prepare,
     onEvent = onEvent
+)
+
+fun <EV : Any, EF : Any> ViewModel.processor(
+    prepare: suspend EllipseContext<Unit, EF>.() -> Unit = {},
+    onEvent: suspend EllipseContext<Unit, EF>.(EV) -> Unit = {},
+): Processor<EV, Unit, EF> = viewModelScope.processor(
+    initialState = Unit,
+    prepare = { prepare(); emptyFlow() },
+    onEvent = { onEvent(it); emptyFlow() }
 )
