+++ ellipse-test/src/main/java/com/tomcz/ellipse/test/ViewModelTestExtension.kt
@@ -2,16 +2,19 @@ package com.tomcz.ellipse.test
 
 import com.tomcz.ellipse.Processor
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.advanceUntilIdle
+import kotlinx.coroutines.test.runCurrent
 import kotlinx.coroutines.test.runTest
 import kotlin.coroutines.CoroutineContext
 
 @ExperimentalCoroutinesApi
-private val defaultAfter: TestScope.() -> Unit = {
-    advanceUntilIdle()
-}
+private val defaultAfter: TestScope.() -> Unit = { advanceUntilIdle() }
+
+@ExperimentalCoroutinesApi
+private val defaultCleanup: TestScope.() -> Unit = {}
 
 @ExperimentalCoroutinesApi
 fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> processorTest(
@@ -23,7 +26,7 @@ fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> processorTest(
     afterEvents: TestScope.() -> Unit = defaultAfter,
     thenStates: TestResultContext<ST>.() -> Unit = {},
     thenEffects: TestResultContext<EF>.() -> Unit = {},
-    cleanup: TestScope.() -> Unit = {}
+    cleanup: TestScope.() -> Unit = defaultCleanup
 ): Unit = processorTest(
     processor = processor,
     context = context,
@@ -46,7 +49,7 @@ fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> processorTest(
     afterEvents: TestScope.() -> Unit = defaultAfter,
     thenStates: TestResultContext<ST>.() -> Unit = {},
     thenEffects: TestResultContext<EF>.() -> Unit = {},
-    cleanup: TestScope.() -> Unit = {}
+    cleanup: TestScope.() -> Unit = defaultCleanup
 ) = runTest(context) {
     processorTest(
         processor = processor,
@@ -60,51 +63,8 @@ fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> processorTest(
     )
 }
 
-// @ExperimentalCoroutinesApi
-// fun <EV : Any, EF : Any, T : Processor<EV, Nothing, EF>> processorTest(
-//     context: CoroutineContext = UnconfinedTestDispatcher(),
-//     processor: TestScope.() -> T,
-//     given: suspend TestScope.() -> Unit = {},
-//     whenEvent: EV,
-//     afterPrepare: TestScope.() -> Unit = defaultAfter,
-//     afterEvents: TestScope.() -> Unit = defaultAfter,
-//     thenEffects: TestResultContext<EF>.() -> Unit = {},
-//     cleanup: TestScope.() -> Unit = {}
-// ): Unit = processorTest(
-//     processor = processor,
-//     context = context,
-//     given = given,
-//     whenEvents = listOf(whenEvent),
-//     afterPrepare = afterPrepare,
-//     afterEvents = afterEvents,
-//     thenEffects = thenEffects,
-//     cleanup = cleanup
-// )
-
-// @ExperimentalCoroutinesApi
-// fun <EV : Any, EF : Any, T : Processor<EV, Nothing, EF>> processorTest(
-//     context: CoroutineContext = UnconfinedTestDispatcher(),
-//     processor: TestScope.() -> T,
-//     given: suspend TestScope.() -> Unit = {},
-//     whenEvents: List<EV> = emptyList(),
-//     afterPrepare: TestScope.() -> Unit = defaultAfter,
-//     afterEvents: TestScope.() -> Unit = defaultAfter,
-//     thenEffects: TestResultContext<EF>.() -> Unit = {},
-//     cleanup: TestScope.() -> Unit = {}
-// ) = runTest(context) {
-//     processorTest(
-//         processor = processor,
-//         given = given,
-//         whenEvents = whenEvents,
-//         afterPrepare = afterPrepare,
-//         afterEvents = afterEvents,
-//         thenEffects = thenEffects,
-//         cleanup = cleanup
-//     )
-// }
-
 @ExperimentalCoroutinesApi
-suspend fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> TestScope.processorTest(
+fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> TestScope.processorTest(
     processor: TestScope.() -> T,
     given: suspend TestScope.() -> Unit = {},
     whenEvent: EV,
@@ -112,7 +72,7 @@ suspend fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> TestScope.
     afterEvents: TestScope.() -> Unit = defaultAfter,
     thenStates: TestResultContext<ST>.() -> Unit = {},
     thenEffects: TestResultContext<EF>.() -> Unit = {},
-    cleanup: TestScope.() -> Unit = {}
+    cleanup: TestScope.() -> Unit = defaultCleanup
 ): Unit = processorTest(
     processor = processor,
     given = given,
@@ -125,7 +85,7 @@ suspend fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> TestScope.
 )
 
 @ExperimentalCoroutinesApi
-suspend fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> TestScope.processorTest(
+fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> TestScope.processorTest(
     processor: TestScope.() -> T,
     given: suspend TestScope.() -> Unit = {},
     whenEvents: List<EV> = emptyList(),
@@ -133,9 +93,10 @@ suspend fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> TestScope.
     afterEvents: TestScope.() -> Unit = defaultAfter,
     thenStates: TestResultContext<ST>.() -> Unit = {},
     thenEffects: TestResultContext<EF>.() -> Unit = {},
-    cleanup: TestScope.() -> Unit = {}
+    cleanup: TestScope.() -> Unit = defaultCleanup
 ) {
-    given()
+    launch { given() }
+    runCurrent()
     val (states, effects) = getStatesAndEffects(
         processorFactory = processor,
         events = whenEvents,
@@ -148,44 +109,3 @@ suspend fun <EV : Any, ST : Any, EF : Any, T : Processor<EV, ST, EF>> TestScope.
     effectsContext.thenEffects()
     cleanup()
 }
-
-// @ExperimentalCoroutinesApi
-// suspend fun <EV : Any, EF : Any, T : Processor<EV, Nothing, EF>> TestScope.processorTest(
-//     processor: TestScope.() -> T,
-//     given: suspend TestScope.() -> Unit = {},
-//     whenEvent: EV,
-//     afterPrepare: TestScope.() -> Unit = defaultAfter,
-//     afterEvents: TestScope.() -> Unit = defaultAfter,
-//     thenEffects: TestResultContext<EF>.() -> Unit = {},
-//     cleanup: TestScope.() -> Unit = {}
-// ): Unit = processorTest(
-//     processor = processor,
-//     given = given,
-//     whenEvents = listOf(whenEvent),
-//     afterPrepare = afterPrepare,
-//     afterEvents = afterEvents,
-//     thenEffects = thenEffects,
-//     cleanup = cleanup,
-// )
-
-// @ExperimentalCoroutinesApi
-// suspend fun <EV : Any, EF : Any, T : Processor<EV, Nothing, EF>> TestScope.processorTest(
-//     processor: TestScope.() -> T,
-//     given: suspend TestScope.() -> Unit = {},
-//     whenEvents: List<EV> = emptyList(),
-//     afterPrepare: TestScope.() -> Unit = defaultAfter,
-//     afterEvents: TestScope.() -> Unit = defaultAfter,
-//     thenEffects: TestResultContext<EF>.() -> Unit = {},
-//     cleanup: TestScope.() -> Unit = {}
-// ) {
-//     given()
-//     val effects = getEffects(
-//         processorFactory = processor,
-//         events = whenEvents,
-//         afterPrepare = afterPrepare,
-//         afterEvents = afterEvents,
-//     )
-//     val effectsContext = TestResultContext(effects, testScheduler)
-//     effectsContext.thenEffects()
-//     cleanup()
-// }
